git-svn-id: http://svn.beutner.name/trunk/sbnc@576 b3a6972a-4407-0410-a3f0-9bc6c1699d69

+++ src/Core.h
@@ -74,7 +74,7 @@ class CCore {
 
 	CVector<CDnsQuery *> m_DnsQueries;
 
-	size_t m_SendqSizeCache;
+	int m_SendqSizeCache;
 
 	SSL_CTX *m_SSLContext;
 	SSL_CTX *m_SSLClientContext;
